# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Demand Planner Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/import/sources": {
      "get": {
        "tags": [
          "import"
        ],
        "summary": "List Sources",
        "operationId": "list_sources_api_v1_import_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceConfigOut"
                  },
                  "type": "array",
                  "title": "Response List Sources Api V1 Import Sources Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Create Source",
        "operationId": "create_source_api_v1_import_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceConfigIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceConfigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/import/sources/{source_id}/import": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Import Data",
        "operationId": "import_data_api_v1_import_sources__source_id__import_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/import/lookup": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Import Lookup",
        "operationId": "import_lookup_api_v1_import_lookup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookupImportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupImportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/forecasts/generate": {
      "post": {
        "tags": [
          "forecasts"
        ],
        "summary": "Generate",
        "operationId": "generate_api_v1_forecasts_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/forecasts/sessions": {
      "get": {
        "tags": [
          "forecasts"
        ],
        "summary": "List Sessions",
        "operationId": "list_sessions_api_v1_forecasts_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionOut"
                  },
                  "type": "array",
                  "title": "Response List Sessions Api V1 Forecasts Sessions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/forecasts/sessions/{session_id}": {
      "get": {
        "tags": [
          "forecasts"
        ],
        "summary": "Get Session",
        "operationId": "get_session_api_v1_forecasts_sessions__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forecasts/sessions/{session_id}/status": {
      "get": {
        "tags": [
          "forecasts"
        ],
        "summary": "Get Status",
        "operationId": "get_status_api_v1_forecasts_sessions__session_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forecasts": {
      "get": {
        "tags": [
          "forecasts"
        ],
        "summary": "List Forecasts",
        "operationId": "list_forecasts_api_v1_forecasts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "product_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ForecastRowOut"
                  },
                  "title": "Response List Forecasts Api V1 Forecasts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forecasts/sessions/{session_id}/publish": {
      "post": {
        "tags": [
          "forecasts"
        ],
        "summary": "Publish",
        "operationId": "publish_api_v1_forecasts_sessions__session_id__publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ForecastRowOut": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "forecast_value": {
            "type": "number",
            "title": "Forecast Value"
          },
          "confidence_lower": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Lower"
          },
          "confidence_upper": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Upper"
          },
          "forecast_type": {
            "type": "string",
            "title": "Forecast Type"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "date",
          "forecast_value",
          "forecast_type"
        ],
        "title": "ForecastRowOut"
      },
      "GenerateIn": {
        "properties": {
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name",
            "default": "default"
          },
          "horizon": {
            "type": "integer",
            "title": "Horizon",
            "default": 28
          },
          "aggregation": {
            "type": "string",
            "title": "Aggregation",
            "default": "monthly"
          },
          "mapping": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Mapping",
            "default": {}
          }
        },
        "type": "object",
        "title": "GenerateIn"
      },
      "GenerateOut": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status"
        ],
        "title": "GenerateOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportIn": {
        "properties": {
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Rows"
          },
          "upload_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Date"
          }
        },
        "type": "object",
        "required": [
          "rows"
        ],
        "title": "ImportIn"
      },
      "ImportOut": {
        "properties": {
          "upload_id": {
            "type": "string",
            "title": "Upload Id"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          }
        },
        "type": "object",
        "required": [
          "upload_id",
          "row_count",
          "columns"
        ],
        "title": "ImportOut"
      },
...
      "LookupImportIn": {
        "properties": {
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Rows"
          },
          "column_mappings": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Column Mappings"
          }
        },
        "type": "object",
        "required": [
          "rows",
          "column_mappings"
        ],
        "title": "LookupImportIn"
      },
      "LookupImportOut": {
        "properties": {
          "inserted": {
            "type": "integer",
            "title": "Inserted"
          }
        },
        "type": "object",
        "required": [
          "inserted"
        ],
        "title": "LookupImportOut"
      },
...
      "SessionOut": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "dataset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Name"
          },
          "horizon": {
            "type": "integer",
            "title": "Horizon"
          },
          "aggregation": {
            "type": "string",
            "title": "Aggregation"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "model_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Used"
          },
          "sku_count": {
            "type": "integer",
            "title": "Sku Count"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "horizon",
          "aggregation",
          "status",
          "sku_count",
          "row_count"
        ],
        "title": "SessionOut"
      },
...
      "SourceConfigIn": {
        "properties": {
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "column_mappings": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Column Mappings",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "file_type"
        ],
        "title": "SourceConfigIn"
      },
      "SourceConfigOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "column_mappings": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Column Mappings"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "file_name",
          "file_type",
          "column_mappings",
          "is_active"
        ],
        "title": "SourceConfigOut"
      },
      "StatusOut": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status"
        ],
        "title": "StatusOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}